+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   - [#392](https://github.com/wix-incubator/rich-content/pull/392) text in line toolbar has tooltips
 - `image`
   - [75336e34](https://github.com/wix-incubator/rich-content/commit/75336e34) open imageStudio on crop tab
+- `gallery`
+  - [#403](https://github.com/wix-incubator/rich-content/pull/403) re-render gallery on change only
 ### :house: Internal
 - `e2e`
   - [#393](https://github.com/wix-incubator/rich-content/pull/393) added atomic block alignment testing